+++ lib/utils/highlighter.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Highlighter {
-  int? index;
-  TextStyle? style;
-  String? word;
-  Highlighter({this.index, this.style, this.word});
-}