+++ lib/utils/word_highlight.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+// A word Highlighter
+class WordHighlight {
+  int? wordIndex;
+  TextStyle? wordStyle;
+  String? word;
+  WordHighlight({this.wordIndex, this.wordStyle, this.word});
+}