+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 0.0.1
+* Describe initial release.
 
-* TODO: Describe initial release.
+## 0.0.2
+* Added common text style
+
+## 0.0.3
+* Documentation added properly.
+* Example added
+* textStyle property added to TextWordHighlighter
+* Highlighter renamed to WordHighlight
+* index inside Highlighter renamed to wordIndex, and style as wordStyle.
+* Some know issues(Font size issues) fixed.
+* Space issue fixed.